requirements.txt
# Core
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=2.0.0
fastapi>=0.110.0
uvicorn>=0.29.0

# PDF parsing
pypdf>=4.0.0
langchain-text-splitters>=0.3.0

# Embeddings + reranking
sentence-transformers>=3.0.0
torch>=2.3.0
transformers>=4.40.0

# Vector index
faiss-cpu>=1.8.0

# Dev / CI
ruff>=0.4.0
pytest>=8.0.0
